+++ apps/api/src/app/core/get-value.object.ts
@@ -1,0 +1,5 @@
+export interface GetValueObject {
+  date: Date;
+  symbol: string;
+  marketPrice: number;
+}
