+++ apps/api/src/app/core/get-value.params.ts
@@ -1,0 +1,8 @@
+import { Currency } from '@prisma/client';
+
+export interface GetValueParams {
+  date: Date;
+  symbol: string;
+  currency: Currency;
+  userCurrency: Currency;
+}
