+++ apps/api/src/app/core/get-values.params.ts
@@ -1,0 +1,9 @@
+import { DateQuery } from '@ghostfolio/api/app/core/market-data.service';
+import { Currency } from '@prisma/client';
+
+export interface GetValuesParams {
+  dateQuery: DateQuery;
+  symbols: string[];
+  currencies: { [symbol: string]: Currency };
+  userCurrency: Currency;
+}
