+++ apps/api/src/app/core/portfolio-calculator.spec.ts
@@ -1,15 +1,7 @@
 import {
-  CurrentRateService,
-  GetValueParams,
-  GetValuesParams
+  CurrentRateService
 } from '@ghostfolio/api/app/core/current-rate.service';
-import {
-  PortfolioCalculator,
-  PortfolioOrder,
-  TimelinePeriod,
-  TimelineSpecification,
-  TransactionPoint
-} from '@ghostfolio/api/app/core/portfolio-calculator';
+import { PortfolioCalculator } from '@ghostfolio/api/app/core/portfolio-calculator';
 import { OrderType } from '@ghostfolio/api/models/order-type';
 import { resetHours } from '@ghostfolio/common/helper';
 import { Currency } from '@prisma/client';
@@ -21,6 +13,12 @@ import {
   isBefore,
   parse
 } from 'date-fns';
+import { PortfolioOrder } from '@ghostfolio/api/app/core/portfolio-order';
+import { TimelinePeriod } from '@ghostfolio/api/app/core/timeline-period';
+import { TimelineSpecification } from '@ghostfolio/api/app/core/timeline-specification';
+import { TransactionPoint } from '@ghostfolio/api/app/core/transaction-point';
+import { GetValuesParams } from '@ghostfolio/api/app/core/get-values.params';
+import { GetValueParams } from '@ghostfolio/api/app/core/get-value.params';
 
 function toYearMonthDay(date: Date) {
   const year = date.getFullYear();
