+++ apps/api/src/app/core/timeline-period.ts
@@ -1,0 +1,8 @@
+import Big from 'big.js';
+
+export interface TimelinePeriod {
+  date: string;
+  grossPerformance: Big;
+  investment: Big;
+  value: Big;
+}
