+++ apps/api/src/app/core/timeline-specification.ts
@@ -1,0 +1,6 @@
+export type Accuracy = 'year' | 'month' | 'day';
+
+export interface TimelineSpecification {
+  start: string;
+  accuracy: Accuracy;
+}
