+++ apps/api/src/app/core/transaction-point.ts
@@ -1,0 +1,6 @@
+import { TransactionPointSymbol } from '@ghostfolio/api/app/core/transaction-point-symbol';
+
+export interface TransactionPoint {
+  date: string;
+  items: TransactionPointSymbol[];
+}
